LostPeterEngine/LostPeterPluginRendererVulkan: Texture

+++ LostPeterVulkan/LostPeterEngine/include/ResourceGroupListener.h
@@ -28,13 +28,13 @@ namespace LostPeterEngine
         
     public:
         //Script
-		virtual void ResourceGroupScriptingStarted(const String& strGroupName, size_t nScriptCount) = 0;	
+		virtual void ResourceGroupScriptingStarted(const String& strGroupName, uint32 nScriptCount) = 0;	
 		virtual void ScriptParseStarted(const String& scriptName, bool& bIsSkipThisScript)	= 0;	
 		virtual void ScriptParseEnded(const String& scriptName, bool bSkipped) = 0;	
 		virtual void ResourceGroupScriptingEnded(const String& strGroupName) = 0;	
 
 		//Resource Load
-		virtual void ResourceGroupLoadStarted(const String& strGroupName, size_t nResourceCount) = 0;	
+		virtual void ResourceGroupLoadStarted(const String& strGroupName, uint32 nResourceCount) = 0;	
 		virtual void ResourceLoadStarted(const Resource* pResource) = 0;	
 		virtual void ResourceLoadEnded() = 0;	
 		virtual void WorldGeometryStageStarted(const String& strDescription) = 0;
@@ -42,7 +42,7 @@ namespace LostPeterEngine
 		virtual void ResourceGroupLoadEnded(const String& strGroupName) = 0;	
 
 		//Resource Prepare
-		virtual void ResourceGroupPrepareStarted(const String& strGroupName, size_t nResourceCount) {}		
+		virtual void ResourceGroupPrepareStarted(const String& strGroupName, uint32 nResourceCount) {}		
 		virtual void ResourcePrepareStarted(const Resource* pResource) {}		
 		virtual void ResourcePrepareEnded() {}		
 		virtual void WorldGeometryStagePrepareStarted(const String& strDescription) {}
